+++ Bjorn/docs/release-notes.md
@@ -2,6 +2,11 @@
 
 # Release notes
 
+## Upcoming release
+
+* Fixed the Lil' Hadron Collider card analyzing non-analyzable cards.
+* Fixed the Lil' Hadron Collider card not counting as analyzing cards for some effects (for example, the Side Projects artifact).
+
 ## 1.0.0
 Released 29 March 2025.